requirements.txt
ansible
appdirs
beautifulsoup4
click
croniter
github3.py
gitpython
influxdb
Jinja2
jenkins-job-builder
jenkinsapi
jira
jmespath
openstacksdk
packaging
python-dateutil
python-novaclient
python-swiftclient
python-subunit
pytz
PyYAML
rackspace-monitoring
rackspace-monitoring-cli
rackspaceauth
requests
sh
shade
six
urllib3
